refactor(core): split interfaces.ts by which side of the engine needs it
interfaces.ts held four unrelated groups in one file, with the build/version
types (BuildInfo, HarnessInfo, VersionReport) physically splitting the
Environment section in half — BackgroundCommandInfo and BackgroundSubagentInfo
sat 100 lines apart with an unrelated block between them.

- interfaces/llm.ts — the model-request contract
- interfaces/environment.ts — tool execution, configuration, and the
  management-plane API (listings, stop/steer, listener attachments)
- interfaces/shared.ts — the vocabulary both genuinely need: ToolDefinition,
  ThinkingLevelName, ApproveFn, the command policy
- interfaces/index.ts — the barrel, still the published
  @prismshadow/penguin-core/interfaces entry
- version-info.ts — build/harness identity, which is neither contract; it now
  ships from the main barrel instead

Pure move: no signature, no behavior and no exported name changed, apart from
the version types leaving the /interfaces subpath for the package root.
