+++ packages/core/src/llm/generative-model.ts
@@ -71,7 +71,7 @@ import type {
   LLMOutcome,
   ThinkingLevelName,
   ToolDefinition,
-} from "../interfaces.js";
+} from "../interfaces/index.js";
 import { ToolCallIdAllocator, stripToolCallIdSuffix } from "./tool-call-ids.js";
 import {
   approximateMessagesTokens,
